requirements.txt
anyio==3.6.2
asgiref==3.6.0
attrs==22.2.0
certifi==2022.12.7
charset-normalizer==3.0.1
click==8.1.3
Django==4.1.7
djangorestframework==3.14.0
dnspython==2.3.0
email-validator==1.3.1
exceptiongroup==1.1.0
fastapi==0.92.0
h11==0.14.0
httpcore==0.16.3
httptools==0.5.0
httpx==0.23.3
idna==3.4
iniconfig==2.0.0
itsdangerous==2.1.2
Jinja2==3.1.2
MarkupSafe==2.1.2
orjson==3.8.6
packaging==23.0
pluggy==1.0.0
pydantic==1.10.5
pytest==7.2.1
pytest-django==4.5.2
pytest-mock==3.10.0
python-dotenv==1.0.0
python-multipart==0.0.5
pytz==2022.7.1
PyYAML==6.0
requests==2.28.2
rfc3986==1.5.0
six==1.16.0
sniffio==1.3.0
sqlparse==0.4.3
starlette==0.25.0
tomli==2.0.1
typing-extensions==4.5.0
ujson==5.7.0
urllib3==1.26.14
uvicorn==0.20.0
uvloop==0.17.0
watchfiles==0.18.1
websockets==10.4
